separate ingredient errors
